+++ changelog.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Improve randomized flag image quality ([#49])
+- Pressing `Enter` key on an exact country name will select the first country in the list corresponding to the name ([#48])
+- Hide the autocomplete when the game is over ([#46])
 
 ## 0.1.0 &#8212; 2024-09-11
 
@@ -23,5 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add randomized flag game
 
 <!-- 0.2.0 -->
+[#46]: https://github.com/wavepulse/atlas/issues/46
+[#48]: https://github.com/wavepulse/atlas/issues/48
 [#49]: https://github.com/wavepulse/atlas/issues/49
 [#50]: https://github.com/wavepulse/atlas/issues/50